pyproject.toml
[project]
name = "commit-card"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.109.0",
    "firebase-admin==6.4.0",
    "gitpython==3.1.41",
    "google-generativeai==0.3.2",
    "httpx==0.26.0",
    "pydantic==2.6.0",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "uvicorn[standard]==0.27.0",
]
